osmflatc: Modularize the relations code
In the original work, nodes and ways were modularized.
This commit does the same for relations.
